fix: move tokenizer methods from user model to a module
