+++ app/models/concerns/tokenizer.rb
@@ -1,0 +1,12 @@
+module Tokenizer
+  def digest(string)
+    cost = ActiveModel::SecurePassword.min_cost ? BCrypt::Engine::MIN_COST :
+                                                   BCrypt::Engine::cost
+
+    BCrypt::Password.create(string, cost: cost)
+  end
+
+  def new_token
+    SecureRandom.urlsafe_base64
+  end
+end
