+++ app/models/user.rb
@@ -1,23 +1,12 @@
 class User < ApplicationRecord
+  extend Tokenizer
+
   attr_accessor :remember_token
 
   validates :username, :email, presence: true, uniqueness: true
   validates :email, email: true
   validates :password, confirmation: true
   has_secure_password
-
-  class << self
-    def digest(string)
-      cost = ActiveModel::SecurePassword.min_cost ? BCrypt::Engine::MIN_COST :
-                                                    BCrypt::Engine::cost
-  
-      BCrypt::Password.create(string, cost: cost)
-    end
-  
-    def new_token
-      SecureRandom.urlsafe_base64
-    end
-  end
 
   def remember
     self.remember_token = User.new_token
